wheel velocity close loop control

+++ software/libs/Inc/encoder.h
@@ -11,6 +11,7 @@
 #include "stm32f4xx_hal.h"
 #include "time_base.h"
 #include <stdint.h>
+#include <math.h>
 
 typedef struct {
 	TIM_HandleTypeDef *timer;
@@ -24,7 +25,7 @@ typedef struct {
 void Encoder_Init(Encoder_t *, TIM_HandleTypeDef *, uint16_t);
 void Encoder_Reset(Encoder_t *);
 
-uint16_t Encoder_GetPositionRaw(Encoder_t *);
+uint32_t Encoder_GetPositionRaw(Encoder_t *);
 
 float Encoder_GetPosition(Encoder_t *);
 float Encoder_GetVelocity(Encoder_t *);
